// src/app/models/admin/capphephoatdongkhoangsan/cpkhaithackhoangsan/cpkhaithacthietbi.model.ts
export class OutputCpKhaiThacThietBiModel {
    idthietbi: string;
    tenthietbi: string;
    mota: string;
    idcapphepkhaithac: string;
    siteid: string;
}

export class InputCpKhaiThacThietBiModel {
    idthietbi: string;
    tenthietbi: string;
    mota: string;
    idcapphepkhaithac: string;
}